feat(recipe): method to multiply ingredients by nr of people

+++ src/main/java/no/ntnu/idatt1005/RecipeList.java
@@ -13,7 +13,7 @@ public class RecipeList {
     return recipeList;
   }
 
-  public void addRecipe(String id, Map<String, Integer> ingredients, String instructions,
+  public void addRecipe(String id, Map<String, Double> ingredients, String instructions,
                         int numberOfPeople) throws IllegalArgumentException {
     if (recipeList.containsKey(id)) {
       throw new IllegalArgumentException("There already exists a recipe with the requested id");
