+++ src/test/java/no/ntnu/idatt1005/TestRecipe.java
@@ -12,9 +12,9 @@ public class TestRecipe {
   @Test
   @DisplayName("Test constructor with valid input")
   void testConstructorWithValidInput() {
-    Map<String, Integer> ingredients = new HashMap<>();
-    ingredients.put("1", 400);
-    ingredients.put("4", 100);
+    Map<String, Double> ingredients = new HashMap<>();
+    ingredients.put("1", 400.0);
+    ingredients.put("4", 100.0);
     String instructions = "Add the chicken nuggets to the pasta";
     int numberOfPeople = 4;
